Display gcc error messages when compilation fails

+++ src/main.rs
@@ -6,7 +6,7 @@ mod tacky;
 
 use std::{ffi::OsStr, io::Write, process::Command};
 
-use anyhow::{bail, ensure, Context, Result};
+use anyhow::{anyhow, bail, ensure, Context, Result};
 use clap::Parser;
 
 use asm::AsmGen;
@@ -171,11 +171,14 @@ fn assemble_and_link(args: &Args, asm: &str) -> Result<()> {
         .output()
         .context("Failed to run gcc for assembly and linking")?;
 
-    ensure!(
-        output.status.success(),
-        "gcc exited with error code {}",
-        output.status,
-    );
+    if !output.status.success() {
+        return Err(anyhow!("gcc exited with error code: {}", output.status)).with_context(|| {
+            format!(
+                "gcc stdout output: \n{}",
+                String::from_utf8_lossy(&output.stderr)
+            )
+        });
+    }
 
     std::fs::remove_file(&asm_path).with_context(|| {
         format!(
